posix: Rework handling of empty path components
Instead of storing empty components, the path decomposition
algorithm now drops them. This ensures that downstream file
systems do not need to take empty components into account
(for example in traverseLinks()).

We still need to handle the case of a trailing slash; we now
do that by setting a Boolean flag. Trailing slashes cause the
path to always be treated as a directory.