posix: implement PRIME for sharing buffers by fd
This commit adds support for the `PRIME_HANDLE_TO_FD` and
`PRIME_FD_TO_HANDLE` ioctls. These convert between a DRM handle (as used
for the MAP_DUMB ioctl) and a fd, which supports the regular operations,
including being sent over a UNIX socket and being `dup`ed. This is
achieved by managing the shared buffers at the device level by being
exported and imported. The shared buffers are identified by the
credentials (a unique identifier) of the passthrough lane servicing the
file operations, as fds are obviously not stable across threads.