posix: Implement killpg(int, int) support
When receiving a kill syscall, check if the PID is less than 0 and if so kill the entire group.
Also, apply the todo logic in the kill(0) path so that the entire group of the running process is killed.

Set errno to ESRCH if the process group is not found, and apply the same logic to other PID and PGRP searches.