drivers/kbd: allow for skipping scancode set selection
When booting on my ThinkPad Yoga 370, booting failed due to multiple
failures that this commit addressed. Firstly, the identifier for the
keyboard was 0xAB84, which has been whitelisted. Secondly, if setting
the scancode fails (which happens on that machine), we just skip it as
we can assume that scancode 2 is the default. For good measure, we
ensure that the keyboard is not disabled.