kernel: Fix broken handling of of libgcc
With the GCC13 release, this hack will become obsolete, because I'll fix
up the -kernel toolchain to not require libc, startfiles or libm at all,
which would permit us to simply not pass -nostdlib instead.  This hinges
on a few outstanding patches that I've already sent upstream, and maybe
an extra ten or so lines of code that would be added locally.

For now, this is good enough.