protocols/fs: make ioctls use a dedicated message format
This changes the way ioctls are represented on the fs protocol. Where
previously they were sent via `CntRequest`, now a two-message format is
used. First, an `IoctlRequest` is sent to identify that the next message
represents an ioctls, as there is no telling whether a standalone
message belongs to an ioctl. Second, the actual message is sent; it is
identified by the message id in the preamble. Currently, all ioctls use
the `GenericIoctlRequest` message; this should be broken up into
per-ioctl messages in the future, but is implemented this way for now to
keep commit size down.