drivers/kbd: increase the PS/2 command timeout to 500 ms
When under very heavy CPU load, my system gets PS/2 command timeouts.
This has also been (very rarely) observed by others, and this is
resolved by increasing the timeout. The latency of this should not
matter, as the boot process is probably already very slow if QEMU is not
getting enough CPU time from the host.