kernel/eir: Merge common aarch64 init code and make virt eir pie
Merge common duplicated aarch64 eir init code and make it a pie on virt platform. It would also likely be possible to do that on raspi4 but I don't really know how the bootloader handles images with base address zero (and if it would be non-zero it would need to be taken into account somehow in eirRelocate() or alternatively making sure there are no relocations at all but that requires some ugly hacks eg. around constexpr string arrays in faults.cpp).