thor: Rework the kernel virtual memory handling
Previously, kernel virtual memory was managed by a buddy tree, which
was problematic, because a lot of larger allocations couldn't be
satisfied, since the buddy allocator ensures that alignment == size.

This commit reworks the kernel VM to instead be managed by a tree of
holes, similar to how user address spaces are managed.

Fixes: #428
Co-authored-by: Alexander van der Grinten <alexander.vandergrinten@gmail.com>