kbd: Handle GetScancodeSet command failing
Some keyboards don't support this command and will just not respond
after ACKing all the command bytes, which makes the command time out.

Linux seems to just assume that if it failed, the scancode set is 2,
so let's do that as well. This seems to make the PS/2 keyboard work on
my laptop.