usb/xhci: Attach completions to all TRBs in chains
Under normal circumstances, the extra events will go unused, because
only the final TRB has interrupt-on-completion set, but in exceptional
situations (e.g. an earlier TRB in the chain failing), previously we
would miss the failure and get permanently stuck (e.g. if the Setup
Stage TRB of a control transfer failed in readDescriptor).

I'm not sure whether this affected short transfers, considering that
they're explicitly chained together via the chain bit, but if it did,
they should be handled correctly now as well.