usb/xhci: Refactor how TRB completion events are tracked
Instead of a single completion attached to a TRB in a ring, now we
have the concept of a transaction, which has a state machine that
manages the events we expect to see for a given type of transfer.

This fixes a few problems, like short transfers accessing the
same (potentially expired) completion twice in certain
cases (multi-TRB chains), and lets us get the transfer size in the
case of short control transfers (although this is not yet exposed to
the USB API).