thor: Generalize CPU timer management logic
Prior to this, multiplexing the preemption and alarm deadline timers
onto a single CPU timer was independently reimplemented on all
architectures.

On AArch64, the logic was changed to always use the virtual timer
instead of both it and the physical timer, and to instead mask the
interrupt instead of setting the deadline to -1 when disarming (which
doesn't work on Apple Silicon apparently).

Fixes #230.