posix: correctly organize class devices in sysfs
Previously, a PCI device that had a class device child `cardX` would
just list it along with its attributes, which was incorrect. Instead,
it should be in a directory with the name of the class (here, `drm`),
where the child device `cardX` should be placed. The `device` link in
`/sys/class/<classname>/<child device name>` correctly points to the
PCI device.