posix: split signal raising logic
By splitting the logic into `determineHandling` and `raiseContext`, we
can correctly raise the cancellation event for signals before they get
raised, while avoid spurious cancellations for e.g. ignored signals.

If a signal is raised while a SignalGuard is active, it is queued as a
delayed signal, which is then raised on `superSigRaise` when the guard
is destructed.

Co-authored-by: Geert Custers <geert.aj.custers@gmail.com>