eir: Use low peripheral mode on Raspberry Pi 4
`dtoverlay=highperi` on it's own does not work fully, and leaves the
GIC in a broken state, since the armstub fails to configure it. Adding
`arm_high_peri=1` to the configuration seems to fix that, but this
isn't documented behavior. High peripheral mode doesn't seem to
provide much benefit anyway (RAM is remapped around the MMIO).