posix: implement POSIX's process vs thread distinction
POSIX defines a process to be a collection of threads. Some properties
are shared (like POSIX timers), while some are per-process (like the
signal mask). This PR does not fully move everything around, but serves
as a starting point to be able to correctly model the relationship.

For now, the interval timers, POSIX timers, parent process, queue of
reapable children at link to /proc/<pid> are owned by the struct
ThreadGroup. More items will need to be moved over, as well as some
logic that needs adjustment. However, this change should not introduce
regressions.